rename jobs/ to graphile-worker/
