+++ plugin-server/src/index.ts
@@ -2,7 +2,7 @@ import { Hub } from '../src/types'
 import { defaultConfig, formatConfigHelp } from './config/config'
 import { healthcheckWithExit } from './healthcheck'
 import { initApp } from './init'
-import { GraphileWorker } from './main/jobs/graphile-worker'
+import { GraphileWorker } from './main/graphile-worker/graphile-worker'
 import { startPluginsServer } from './main/pluginsServer'
 import { Status } from './utils/status'
 import { makePiscina } from './worker/piscina'
