+++ plugin-server/src/main/graphile-worker/worker-setup.ts
