+++ plugin-server/tests/main/capabilities.test.ts
@@ -1,8 +1,8 @@
 import Piscina from '@posthog/piscina'
 
+import { startGraphileWorker } from '../../src/main/graphile-worker/worker-setup'
 import { KafkaQueue } from '../../src/main/ingestion-queues/kafka-queue'
 import { startQueues } from '../../src/main/ingestion-queues/queue'
-import { startGraphileWorker } from '../../src/main/jobs/worker-setup'
 import { Hub, LogLevel } from '../../src/types'
 import { createHub } from '../../src/utils/db/hub'
